% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xs.R
\name{xs_calc_geom}
\alias{xs_calc_geom}
\title{Calculate cross section hydraulic geometry properties}
\usage{
xs_calc_geom(xs, water_surface_elevation)
}
\arguments{
\item{xs}{A tbl_df containing cross section geometry, as returned by the \code{xs_prep} function}

\item{water_surface_elevation}{The water surface elevation at which the properties are calculated}
}
\description{
Input a cross section data frame as returned by the \code{xs_prep} function.
Outputs a named vector of hydraulic parameters (such as cross-sectional area and wetted perimeter) at the specified water surface elevation.
}
