% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xs.R
\name{xs_eval_all}
\alias{xs_eval_all}
\title{Calculate hydraulic geometry properties, WSE, and velocity at multiple discharges}
\usage{
xs_eval_all(xs, rc, discharges)
}
\arguments{
\item{xs}{A \code{tbl_df} containing cross section geometry, as returned by the \code{xs_prep} function}

\item{rc}{A \code{tbl_df} containing a depth-discharge rating curve, as returned by the \code{xs_rating_curve} function}

\item{discharges}{A discharge (cfs) number at which to determine the water surface elevation, or a vector of multiple discharges. Vector can be named or unnamed. Also accepts a data frame or tibble containing a column called "discharge"}
}
\description{
This function takes a cross section data frame as returned by the \code{xs_prep} function, and a rating curve calculated by \code{xs_rating_curve.} It returns hydraulic parameters for one or more given discharges.
}
