% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xs.R
\name{xs_eval_over_rc}
\alias{xs_eval_over_rc}
\title{Calculate hydraulic geometry properties, WSE, and velocity across all discharges in a rating curve}
\usage{
xs_eval_over_rc(xs, rc, sediment_transport = FALSE, slope = NA)
}
\arguments{
\item{xs}{A \code{tbl_df} containing cross section geometry, as returned by the \code{xs_prep} function}

\item{rc}{A \code{tbl_df} containing a depth-discharge rating curve, as returned by the \code{xs_rating_curve} function}

\item{sediment_transport}{Specify whether to calculate bed mobilization and grain suspension. Defaults to FALSE}

\item{slope}{If calculating sediment transport, specify a slope to use for the calculation (normal depth assumption)}
}
\description{
This function takes a cross section data frame as returned by the \code{xs_prep} function, and a rating curve calculated by \code{xs_rating_curve.} It returns hydraulic parameters for all discharges in the provided rating curve
}
