% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xs.R
\name{xs_mannings_n}
\alias{xs_mannings_n}
\title{Back-calculate Manning's n given known slope}
\usage{
xs_mannings_n(xs, water_surface_elev, slope, discharge)
}
\arguments{
\item{xs}{A tbl_df containing cross section geometry, as returned by the `xs_prep` function}

\item{water_surface_elev}{The water surface elevation at which the properties are calculated}

\item{slope}{The channel profile slope at the cross section, used in Manning's equation (normal depth assumption)}

\item{discharge}{A known discharge value in cfs}
}
\description{
Input a cross section data frame as returned by the `xs_prep` function, along with a WSE and slope as might be gathered in a high water mark survey, and a known discharge having produced the given water surface elevation.
Returns a Manning's roughness coefficient by back-calculating from Manning's equation.
}
