% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xs.R
\name{xs_plot}
\alias{xs_plot}
\title{Plot cross section}
\usage{
xs_plot(xs, wse = NA)
}
\arguments{
\item{xs}{A tbl_df containing cross section geometry, as returned by the \code{xs_prep} function}

\item{wse}{A water surface elevation, as returned by the \code{xs_rc_interpolate} function}
}
\description{
This function takes a cross section data frame as returned by the \code{xs_prep} function, and plots it with 1:1 scale. Optionally also adds a given water surface line.
}
