% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xs.R
\name{xs_plot2}
\alias{xs_plot2}
\title{Plot two cross sections}
\usage{
xs_plot2(
  xs1,
  xs2,
  wse1 = NA,
  wse2 = NA,
  label1 = "Baseline",
  label2 = "Design"
)
}
\arguments{
\item{xs1}{A tbl_df containing a cross section geometry, as returned by the \code{xs_prep} function}

\item{xs2}{A tbl_df containing a cross section geometry, as returned by the \code{xs_prep} function}

\item{wse1}{A water surface elevation for \code{xs1}, as returned by the \code{xs_rc_interpolate} function}

\item{wse2}{A water surface elevation for \code{xs2}, as returned by the \code{xs_rc_interpolate} function}

\item{label1}{The label for the first cross section \code{xs1}. Defaults to "Baseline"}

\item{label2}{The label for the second cross section \code{xs2}. Defaults to "Design"}
}
\description{
This function takes two cross section datas frame as returned by the \code{xs_prep} function, and plots them with 1:1 scale. Optionally also adds a given water surface line.
}
