% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xs.R
\name{xs_plot_rc}
\alias{xs_plot_rc}
\title{Plot rating curve}
\usage{
xs_plot_rc(
  rc,
  y = water_surface_elevation,
  x = discharge,
  ylab = "Water Surface Elevation (ft)",
  xlab = "Discharge (cfs)"
)
}
\arguments{
\item{rc}{A \code{tbl_df} containing a depth-discharge rating curve, as returned by the xs_rating_curve function}

\item{y}{The unquoted name of a rating curve variable to plot on the y axis. Defaults to \code{water_surface_elevation}.}

\item{x}{The unquoted name of a rating curve variable to plot on the x axis. Defaults to \code{discharge}.}

\item{ylab}{If a custom y axis variable is selected, provide a text label for the y axis.}

\item{xlab}{If a custom x axis variable is selected, provide a text label for the x axis.}
}
\description{
This function plots a rating curve as returned by the \code{xs_rating_curve} function.
}
