% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xs.R
\name{xs_plot_sediment}
\alias{xs_plot_sediment}
\title{Plot sediment transport rating curve}
\usage{
xs_plot_sediment(xs, rc, slope)
}
\arguments{
\item{xs}{A tbl_df containing a cross section geometry, as returned by the \code{xs_prep} function}

\item{rc}{A \code{tbl_df} containing a depth-discharge rating curve, as returned by the xs_rating_curve function}

\item{slope}{Specify a slope to use for the calculation (normal depth assumption)}
}
\description{
This function plots the sediment transport rating curves showing grain size mobilized and suspended.
}
