% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xs.R
\name{xs_plot_sediment2}
\alias{xs_plot_sediment2}
\title{Plot two sediment transport rating curves}
\usage{
xs_plot_sediment2(
  xs1,
  rc1,
  xs2,
  rc2,
  label1 = "Baseline",
  label2 = "Design",
  slope
)
}
\arguments{
\item{xs1}{A tbl_df containing a cross section geometry, as returned by the \code{xs_prep} function}

\item{rc1}{A \code{tbl_df} containing a depth-discharge rating curve, as returned by the xs_rating_curve function}

\item{xs2}{A tbl_df containing a cross section geometry, as returned by the \code{xs_prep} function}

\item{rc2}{A \code{tbl_df} containing a depth-discharge rating curve, as returned by the xs_rating_curve function}

\item{label1}{The label for the first cross section \code{xs1}. Defaults to "Baseline"}

\item{label2}{The label for the second cross section \code{xs2}. Defaults to "Design"}

\item{slope}{If calculating sediment transport, specify a slope to use for the calculation (normal depth assumption)}
}
\description{
This function plots two sediment transport rating curves showing grain size mobilized and suspended.
}
