% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xs.R
\name{xs_rating_curve}
\alias{xs_rating_curve}
\title{Create rating curve for cross section with given slope and roughness}
\usage{
xs_rating_curve(xs, slope, mannings_n, delta_z = 0.1)
}
\arguments{
\item{xs}{A \code{tbl_df} containing cross section geometry, as returned by the \code{xs_prep} function}

\item{slope}{The channel profile slope at the cross section, used in Manning's equation (normal depth assumption)}

\item{mannings_n}{The roughness coefficient for the cross section, used in Manning's equation.}

\item{delta_z}{The elevation interval to be used for calculating outputs at different water surface elevations. Defaults to 0.1 ft.}
}
\description{
This function runs the \code{xs_calc_geom} function along a series of water surface elevations to return a rating curve of water surface elevation versus cross-sectional area and wetted perimeter. Then it applies Manning's equation to estimate depth and velocity. Returns a \code{tbl_df} with one row per water surface elevation.
}
