% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xs.R
\name{xs_rc_interpolate}
\alias{xs_rc_interpolate}
\title{Interpolate WSE from rating curve at given discharge}
\usage{
xs_rc_interpolate(rc, discharge)
}
\arguments{
\item{rc}{A \code{tbl_df} containing a depth-discharge rating curve, as returned by the \code{xs_rating_curve} function}

\item{discharge}{A discharge (cfs) number at which to determine the water surface elevation}
}
\description{
This function takes a rating curve calculated by xs_rating_curve and returns the water surface elevation for a given discharge.
}
