% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xs.R
\name{xs_sediment_transport}
\alias{xs_sediment_transport}
\title{Calculate cross section hydraulic geometry properties and sediment transport estimates}
\usage{
xs_sediment_transport(xs, water_surface_elev, slope)
}
\arguments{
\item{xs}{A tbl_df containing cross section geometry, as returned by the \code{xs_prep} function}

\item{water_surface_elev}{The water surface elevation at which the properties are calculated}

\item{slope}{The channel profile slope at the cross section, used in Shield's equation and estimation of shear velocity (normal depth assumption)}
}
\description{
Input a cross section data frame as returned by the \code{xs_prep} function.
Outputs a named vector with estimates for median mobilized bed grain size and maximum suspended load grain size
}
