% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xsmatic-package.R
\docType{package}
\name{xsmatic-package}
\alias{xsmatic}
\alias{xsmatic-package}
\title{xsmatic: Cross Section Geometry and Manning's Equation Calculations}
\description{
Description here
}
\author{
\strong{Maintainer}: Skyler Lewis \email{slewis@flowwest.com}

}
\keyword{internal}
